+++ clang-tools-extra/clang-tidy/misc/UniqueptrResetReleaseCheck.h
@@ -39,7 +39,6 @@ public:
                            Preprocessor *ModuleExpanderPP) override;
   void registerMatchers(ast_matchers::MatchFinder *Finder) override;
   void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
-  void storeOptions(ClangTidyOptions::OptionMap &Opts) override;
 
 private:
   utils::IncludeInserter Inserter;